feat: add scripts to work with production networks (#59)
## High Level Overview of Change

This PR adds support for working with external networks (namely,
building a bridge, transferring across that bridge, and creating new
accounts across that bridge).

There are also some breaking changes with the methods, because it turns
out there's a better way to pass in the information.

### Context of Change

While it's nice and useful to build with local sidechains, most people
will be working with external networks.

### Type of Change

- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
## Test Plan

Added tests. CI passes.