fix: switch tests to ubuntu 20.04 (#89)
## High Level Overview of Change

This PR switches the Github Actions from running on `ubuntu-latest`
(which recently switched to 22.04) to `ubuntu-20.04`, which fixes tests.

### Context of Change

Ubuntu 22.04 upgraded OpenSSL to version 3.0, which deprecated
ripemd160. Github will not add support to the runner, because they want
to run only the default version.

https://github.com/openssl/openssl/issues/16994
https://github.com/actions/runner-images/issues/6676

### Type of Change

- [x] Tests (You added tests for code that already exists, or your new
feature included in this PR)

## Test Plan

CI now passes.