fix: better handling of long start-up times (#125)
## High Level Overview of Change

This PR handles long start-up times for the servers better. It checks if
the process is a zombie and errors if that occurs (resulting in faster
crashes on errors), and otherwise waits for 30 seconds instead of 10.

### Context of Change

Sometimes, after the server has just been built, it takes longer than
the assigned 10 seconds to start up.

### Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

Works locally. CI passes locally.