fix: improve build flow for production-style networks (#126)
## High Level Overview of Change

This PR refactors the `bridge build` command to (for an XRP-XRP bridge)
transfer the funds to create the issuing chain witness submission
accounts via a `Payment` from the issuing chain's genesis account, and
then transfer the same amount of funds from the funding seed to the
locking chain's door account (to ensure that the amount of funds locked
in the locking chain's door account is the same as the amount of funds
that have been issued on the issuing chain). This avoids using the
bridge for the account creation, which simplifies a lot of issues.

It also adds a check to ensure that when running on an external network,
the user is using `--no-close-ledgers` (it errors otherwise).

### Context of Change

These issues occurred while setting up production-style networks.

### Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

Tests pass locally. The docker container still hasn't been fixed.