change: skip already-done steps when bridge building (#131)
## High Level Overview of Change

This PR enhances the `bridge build` command to skip the steps of the
bridge build that have already been completed (e.g. don't run
`XChainCreateBridge` if the `Bridge` ledger object already exists).

### Context of Change

If you mess up a bridge setup for some reason, you shouldn't be unable
to build a bridge - it makes way more sense to just skip the steps that
have already been done. It's hard to work with in a production-style
network otherwise.

### Type of Change

- [x] New feature (non-breaking change which adds functionality)

## Test Plan

Works locally. Tests pass locally.